WIP Updating loading logic

+++ CsPlugin/Source/CsCore/Types/CsTypes_Macro.h
@@ -18,6 +18,6 @@
 #define CS_SET_BLUEPRINT_BITFLAG(Bitmask, Bit) (Bitmask |= 1 << static_cast<uint32>(Bit))
 #define CS_CLEAR_BLUEPRINT_BITFLAG(Bitmask, Bit) (Bitmask &= ~(1 << static_cast<uint32>(Bit)))
 
-#define CS_TEST_BITFLAG(Bitmask, Bit) ((Bitmask & Bit) == Bit)
-#define CS_SET_BITFLAG(Bitmask, Bit) (Bitmask |= Bit)
-#define CS_CLEAR_BITFLAG(Bitmask, Bit) (Bitmask &= ~Bit)
+#define CS_TEST_BITFLAG(Bitmask, Bit) ((Bitmask & static_cast<uint32>(Bit)) == static_cast<uint32>(Bit))
+#define CS_SET_BITFLAG(Bitmask, Bit) (Bitmask |= static_cast<uint32>(Bit))
+#define CS_CLEAR_BITFLAG(Bitmask, Bit) (Bitmask &= ~static_cast<uint32>(Bit))
